feat: enhance smart-commerce platform with improved mock generation and notification system
Major enhancements:
- Fixed transaction ID collision by switching to timestamp+random format (txn_{ms}_{random})
- Implemented realistic mock data patterns for both restock and survival modes
  * Restock mode: varied intervals between transactions, auto stock decrement
  * Survival mode: strategic backdating (1, 45, 95 days) to ensure threshold triggers
- Enhanced frontend debugging with extensive console logging in Prediksi.jsx
- Updated documentation for all modules including API references
- Added auto-prediction workflow integration from mock to notifications
- Improved customer visibility logic (created_by_store_id or has transactions)
- Fixed store creation and user management issues

Technical details:
- Transaction ID generator replaced sequential pattern to prevent 70+ insert collisions
- Stock synchronization ensures predictions are meaningful after mock seeding
- Survival spacing calculated to meet pred_days_left <= 3 notification threshold
- Comprehensive logging added for easier troubleshooting of click events and state changes

Documentation updates:
- README.md enhanced with architecture overview, workflows, and testing guides
- Backend API docs updated with recent enhancements and configuration examples
- Added section on mock data generation strategies and their ML training benefits

Testing improvements:
- Verification scripts for end-to-end flows
- Mock → Prediction → Notification automated pipeline
- Cross-module integration tests pass successfully

+++ internal/pkg/mlclient/client.go
@@ -48,7 +48,7 @@ type InventoryPredictionBatchRequest struct {
 
 // SurvivalPredictionRequest maps to POST /predict-survival
 type SurvivalPredictionRequest struct {
-	CustomerID       int     `json:"customer_id"`
+	CustomerID       int64   `json:"customer_id"`
 	StockCode        string  `json:"stock_code"`
 	Quantity         float64 `json:"quantity"`
 	UnitPrice        float64 `json:"unit_price"`
@@ -66,7 +66,7 @@ type SurvivalPredictionRequest struct {
 
 // SurvivalPredictionResponse maps to the ML service response
 type SurvivalPredictionResponse struct {
-	CustomerID             int     `json:"customer_id"`
+	CustomerID             int64   `json:"customer_id"`
 	StockCode              string  `json:"stock_code"`
 	PredictedRestockDate   string  `json:"predicted_restock_date"`
 	PredDaysLeft           int     `json:"pred_days_left"`
